refactor(core): split coverage facade
